+++ src/test/extension.test.ts
@@ -1,15 +1,26 @@
-import * as assert from 'assert';
+import * as assert from "assert";
 
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
-import * as vscode from 'vscode';
+import * as vscode from "vscode";
+import * as utils from "../utils";
 // import * as myExtension from '../../extension';
 
-suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
+suite("Extension Test Suite", () => {
+  vscode.window.showInformationMessage("Start all tests.");
 
-	test('Sample test', () => {
-		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
-	});
+  test("C2U", () => {
+    const exampleText = "啊啊ff啊啊🀄ww🀄1122wwww";
+    const genCommand = (
+      includeSurrogate: boolean,
+      ignoreAscii: boolean,
+      utf8: boolean
+    ) => {
+      return utils.toUnicode(
+        utils.toUnicodeArray(exampleText, includeSurrogate),
+        ignoreAscii,
+        utf8
+      );
+    };
+  });
 });