+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.1.1]
+
+- Shrinked the extension size
+- Added a command to convert characters between characters and escaped sequences
+- Added a command to encode to UTF-8 sequences
+
 ## [0.1.0]
 
 - Added configurable underline